rename module to fix imports
