+++ .vscode/settings.json
@@ -5,5 +5,6 @@
     "python.formatting.provider": "black",
     "python.linting.enabled": true,
     "python.linting.pylintEnabled": true,
-    "python.linting.flake8Enabled": true
+    "python.linting.flake8Enabled": true,
+    "pylint.importStrategy": "fromEnvironment"
 }
